src/agents/screenplay_agent.py
"""Screenplay writing agent using Google ADK."""

from typing import Any

from google.adk import Agent
from pydantic import BaseModel


class ScreenplayOutline(BaseModel):
    """Screenplay outline structure."""

    title: str
    logline: str
    three_act_structure: dict[str, str]
    characters: list[dict[str, str]]


class ScreenplayAgent:
    """Agent specialized in creating screenplay outlines."""

    def __init__(self, model: str = "gemini-2.5-flash") -> None:
        """Initialize screenplay agent.

        Args:
            model: LLM model to use
        """
        self.model = model
        self.agent = self._create_agent()

    def _create_agent(self) -> Agent:
        """Create the ADK agent with instructions.

        Returns:
            Configured ADK Agent
        """
        return Agent(
            name="screenplay_writer",
            model=self.model,
            description="Expert in crafting compelling screenplay outlines",
            instruction="""
            You are a professional screenwriter specializing in story structure.

            When given a concept, create a screenplay outline with:
            1. A compelling title
            2. A one-line logline (max 50 words)
            3. Three-act structure breakdown:
               - Act 1: Setup (introduce characters, world, conflict)
               - Act 2: Confrontation (escalate conflict, character development)
               - Act 3: Resolution (climax and resolution)
            4. Main characters with brief descriptions

            Focus on:
            - Clear story arc
            - Emotional resonance
            - Commercial viability
            - Character motivations
            """,
        )

    async def create_outline(self, concept: str) -> dict[str, Any]:
        """Create screenplay outline from concept.

        Args:
            concept: Movie concept or premise

        Returns:
            Screenplay outline data
        """
        # This would use the ADK agent's generate method
        # For now, returning structure for testing
        return {
            "title": "Generated Title",
            "logline": "A compelling story about...",
            "three_act_structure": {
                "act_1": "Setup...",
                "act_2": "Confrontation...",
                "act_3": "Resolution...",
            },
            "characters": [
                {"name": "Protagonist", "description": "Main character"},
            ],
        }
